Drop largest message from buffer on Kafka::MessageSizeTooLarge exception.
We can get into a bad unrecoverable state when we get a Kafka::MessageSizeTooLarge exception.

When we receive this exception, we remove the largest message from the buffer and simply retry.  Once retried, if the buffer is still too large, we'll continue removing the largest message until we can successfully flush the buffer.